+++ cdk/lib/echo-service.ts
@@ -4,13 +4,55 @@ import * as logs from 'aws-cdk-lib/aws-logs';
 import * as ssm from 'aws-cdk-lib/aws-ssm';
 import * as ecs_patterns from 'aws-cdk-lib/aws-ecs-patterns';
 
-import { echoImage, echoOtelImage } from './docker-images';
+import { echoImage, echoImageXray } from './docker-images';
 import { allPorts } from './all-ports';
 import { addEcsRole } from './add-ecs-role'
 import { FargateWithOtelCollectorTaskDefinition } from './fargate-with-otel-collector-task-definition';
 
+// ECHO SERVICE - X-RAY
 
-// ECHO SERVICE
+export function createEchoXrayService(stack: cdk.Stack, cluster: ecs.Cluster, logGroup: logs.LogGroup, swParam: ssm.StringParameter) :  ecs.FargateService {
+  const echoTaskDefinitionXray = new FargateWithOtelCollectorTaskDefinition(stack, 'EchoTaskDefinitionXray', {
+    memoryLimitMiB: 512,
+    cpu: 256,
+    executionRole: addEcsRole(stack, 'addEcsEchoRoleXray')
+  });
+
+  const containerXrayEcho = echoTaskDefinitionXray.addContainer('EchoXrayContainer', {
+    image: echoImageXray,
+    portMappings: [{ containerPort: 8080 }],
+    logging: new ecs.AwsLogDriver({
+      logGroup: logGroup,
+      streamPrefix: 'echo-xray',
+      mode: ecs.AwsLogDriverMode.NON_BLOCKING }),
+    environment: {
+      'WORKERS': '5',
+      // 'AWS_XRAY_CONTEXT_MISSING': 'LOG_ERROR',
+      // 'AWS_XRAY_DAEMON_ADDRESS': 'xray-daemon:2000',
+    },
+  });
+
+  const containerXrayCloudWatchAgent = echoTaskDefinitionXray.addCloudWatchAgentContainer(logGroup, swParam);
+  const containerXRaysidecar = echoTaskDefinitionXray.addXRaysidecarContainer(logGroup);
+
+  const echoServiceXray = new ecs_patterns.NetworkLoadBalancedFargateService(stack, 'EchoServiceXray', {
+    cluster,
+    taskDefinition: echoTaskDefinitionXray,
+    desiredCount: 3,
+    cloudMapOptions: {
+      name: 'echo',
+    },
+    publicLoadBalancer: true,
+    listenerPort: 8080,
+  });
+
+  echoServiceXray.service.connections.allowFromAnyIpv4(allPorts);
+
+  return echoServiceXray.service;
+}
+
+
+// ECHO SERVICE - OTEL
 export function createEchoService(stack: cdk.Stack, cluster: ecs.Cluster, logGroup: logs.LogGroup, swParam: ssm.StringParameter) : ecs.FargateService {
 
   const echoTaskDefinition = new FargateWithOtelCollectorTaskDefinition(stack, 'EchoTaskDefinition', {
@@ -28,8 +70,9 @@ export function createEchoService(stack: cdk.Stack, cluster: ecs.Cluster, logGro
       mode: ecs.AwsLogDriverMode.NON_BLOCKING }),
     environment: {
       'WORKERS': '5',
-      'AWS_XRAY_TRACING_NAME': 'web-service',
-      // 'OTEL_RESOURCE_ATTRIBUTES': 'aws.log.group.names=' + logGroup.logGroupName,
+      // 'AWS_XRAY_TRACING_NAME': 'web-service',
+      'OTEL_SERVICE_NAME': 'echo-service',
+      'OTEL_RESOURCE_ATTRIBUTES': 'aws.log.group.names=' + logGroup.logGroupName,
     },
   });
 
@@ -52,46 +95,4 @@ export function createEchoService(stack: cdk.Stack, cluster: ecs.Cluster, logGro
   echoService.service.connections.allowFromAnyIpv4(allPorts);
 
   return echoService.service;
-}
-
-
-// ECHO OTEL SERVICE
-
-export function createEchoOtelService(stack: cdk.Stack, cluster: ecs.Cluster, logGroup: logs.LogGroup, swParam: ssm.StringParameter) : ecs.FargateService {
-  const echoOtelTaskDefinition = new FargateWithOtelCollectorTaskDefinition(stack, 'EchoOtelTaskDefinition', {
-    memoryLimitMiB: 512,
-    cpu: 256,
-    executionRole: addEcsRole(stack, 'addEcsEchoOtelRole')
-  });
-
-  const containerOtelEcho = echoOtelTaskDefinition.addContainer('EchoOtelContainer', {
-    image: echoOtelImage,
-    portMappings: [{ containerPort: 8080 }],
-    logging: new ecs.AwsLogDriver({
-      logGroup: logGroup,
-      streamPrefix: 'echo-otel',
-      mode: ecs.AwsLogDriverMode.NON_BLOCKING }),
-    environment: {
-      'WORKERS': '5',
-      'OTEL_SERVICE_NAME': 'echo-service-otel',
-      'OTEL_RESOURCE_ATTRIBUTES': 'aws.log.group.names=' + logGroup.logGroupName,
-    },
-  });
-
-  const containerOtelCloudWatchAgent = echoOtelTaskDefinition.addCloudWatchAgentContainer(logGroup, swParam);
-
-  const echoOtelService = new ecs_patterns.NetworkLoadBalancedFargateService(stack, 'EchoOtelService', {
-    cluster,
-    taskDefinition: echoOtelTaskDefinition,
-    desiredCount: 3,
-    cloudMapOptions: {
-      name: 'echo_otel',
-    },
-    publicLoadBalancer: true,
-    listenerPort: 8080,
-  });
-
-  echoOtelService.service.connections.allowFromAnyIpv4(allPorts);
-
-  return echoOtelService.service;
 }
