--- server/web_rubyrails/Dockerfile.rails_base_otel
@@ -1,21 +1,0 @@
-FROM ruby_base AS rails_base
-
-RUN gem install rails -v "6.1.7"
-
-RUN rails new myapp -d sqlite3 --api
-
-WORKDIR /myapp
-
-RUN rake db:create
-
-RUN bundle add opentelemetry-sdk opentelemetry-instrumentation-all opentelemetry-exporter-otlp opentelemetry-propagator-xray
-RUN bundle install
-
-ADD application_controller_otel.rb app/controllers/application_controller.rb
-
-ADD routes.rb config/
-ADD opentelemetry.rb config/initializers/
-
-EXPOSE 8080
-
-CMD ["rails", "server", "-b", "0.0.0.0", "-p", "8080"]
